move evaluation domains into separate library
