+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "dlog",
     "pairing",
     "oracle",
+    "evaluation_domains",
 ]
 
 [profile.release]
