+++ pairing/circuits/Cargo.toml
@@ -15,3 +15,4 @@ rand_core = { version = "0.5" }
 colored = "1.9.2"
 rand = "0.7.3"
 sprs = "0.7.1"
+evaluation_domains = { path = "../../evaluation_domains" }
